IRMG aesthetic updates: the adminbus fun toys (#2281)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
An update to our favorite boys in brown, this time for the higher ups in
the faction and outfits for Admins to wear for events regarding the
IRMG.
Also tacks on a peaked cap for normal Vanguards, not anywhere just yet.

Spin around of the Honorable Vanguard outfit
![honorable
spin](https://github.com/shiptest-ss13/Shiptest/assets/81479835/2d4f6ee1-fc88-4bfa-8daa-6fb60a66e02c)

Yes there's birdie versions but doing spin arounds is pain.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Adminbus :)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
imageadd: Inteq Honorable Vanguard clothing have been added for admins
to play with, along with a peaked cap for normal Vanguards, not yet
mapped anywhere yet.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->