Mudskipper-class Salvage Clipper (#2224)
## About The Pull Request

![mudskipper-2](https://github.com/shiptest-ss13/Shiptest/assets/118859017/5424d514-95fc-4242-9846-9dc06a6e9d4f)

![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/b94454f3-db85-462e-b5fe-db320ee87332)

![mudskipper-area-2](https://github.com/shiptest-ss13/Shiptest/assets/118859017/1e9c04ac-3184-4475-a950-5fba80a374a9)

adds the mudskipper-class salvage clipper, an independent replacement to
the scav-class. the mudskipper is a small, budget scavenger vessel.
housing only 4 crewmembers, it specializes heavily in ruin-raiding &
salvage, at the cost of lacking in every other department. it also
boasts a much lower depressurize-ability compared to the original
scav-class. yippee

## Why It's Good For The Game

the scav-class was an egregious excuse of a ship. this PR will banish it
to The Pit.

## Changelog

:cl:
add: the mudskipper-class salvage clipper, the scavenging successor to
the scav.
del: removed the scav-class
/:cl: