New Lavaland Ruin: The Nanotrasen Genetics Research Facility (#2047)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the abandoned Nanotrasen Genetics Research Facility, a facility on
lavaland dedicated to genetic and slime extract related experiments away
from the prying eyes of the general public. My hope for this ruin is to
act as a Nanotrasen counterpart to the Syndicate bioweapons facility on
lavaland, allowing players from either faction to loot either facility.

The facility is equipped with a Password Locked Door separating the
reception and subject area from the storage and research area, which can
only be unlocked with a code found somewhere within the ruin (hint
hint).

Threats:
7x Hungry Grey Slimes (3 of which are in locked rooms)
2x Powered Turrets (On stun by default, mainly to act as a stealth
puzzle for players to dodge past)

Items of Value: 
3x Green Slime Extract
1x Dark Blue Slime Extract
1x Healing Extract
1x Invigorating Extract
4x Cryoxadone Beakers
2x Cryogenic Cell setup (If they even bother to deconstruct it and take
it)
1x Medical Sleeper
5x RTG Generator
3x Fashionable Ties
1120 Credits
As well as whatever various plushies, junk and syringes you manage to
loot from the storage and research area.

![2023-07-18 21 25
16](https://github.com/shiptest-ss13/Shiptest/assets/95449138/f8ccbdd7-8d55-414f-8429-a1588d95f7e3)


![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/8d29a753-e5f5-4b2e-b5f1-9e5bec54519d)



## Why It's Good For The Game
I was asked to make new, unique lavaland ruins due to most of them being
kind of shit and soon to be deprecated. This will be the first of many
new ruins for lavaland.
This is also the first non-deprecated ruin to use the password locked
door, something I've been wanting to use.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Nanotrasen Genetic Research Facility, a new Lavaland ruin
add: codelab.dm containing assets for codelab.dmm
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>