Total Remap of the Ranger (#2034)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
## About The Pull Request
The old ranger is just about a year old. This totally remaps this ship
and rebalances it for our current standards.

![rangerpng](https://github.com/shiptest-ss13/Shiptest/assets/55075747/1290c0e9-5efb-4235-b047-a8993466f5ca)
![2023-08-19 12 35
04](https://github.com/shiptest-ss13/Shiptest/assets/55075747/acfb106e-b770-4f9a-8f76-bdbcb3dee6ff)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
This is good for the game because it replaces a year old ship with
something that's more up to par with our standards.
## Why It's Good For The Game
Ship upkeep and maintaining our maps is a good thing. Plus the old
ranger had a bunch of old snowflaked code that was a bandaid fix for the
ship access system not existing yet. (It unsnowflakes it).
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Remaps the Ranger-class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>