IRMG Vaquero-class Light Frigate (#2214)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![vaquero](https://github.com/shiptest-ss13/Shiptest/assets/60533805/cb3c81e9-9875-4d45-92d5-0bd859f79bc4)

![vaquero_areas](https://github.com/shiptest-ss13/Shiptest/assets/60533805/a23bae53-854b-4f0d-9182-bee25394e24c)

Adds the Vaquero-class Light Frigate, a small (7 players) Inteq
generalist to fill the frankly terrible shoes the Hound left behind.
While the Vaquero has all the features expected of an Inteq ship
(meeting room, armory, large cargo bay, boombox), it actually has
average engineering gear in combination with a full set of surgical
tools and table, making it much more well-rounded than the other Inteq
ships, even if it pays for this by only being moderately armed and
having a small footprint and crew capacity.

The Vaquero has a crew capacity of 7:
- one Vanguard
- one Master-At-Arms to run the armory and function as second-in-command
- one Enforcer
- one Artificer
- one Corpsman
- two Recruits

## Why It's Good For The Game

A generalist rounds out the Inteq catalogue nicely, and helps balance
out the fact that all of the Inteq specialists are positively enormous.

## Changelog

:cl:
add: Added the IRMG Vaquero
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>