Pill Press (#2275)
## Pill Press!
This will add a lacking feature in ghettochem, the ability to produce
pills.

### todo

- [x] Make it work!
- [x] Figure out how to select what pill type you want to use, when
altclick, screwdriver, and wrench are all taken!
- [x] Learn how overlays work!
- [x] Sprites!
- [x] Get yelled at by coders probably!
- [x] Re-path it to be more in-line with other chem devices
- [x] Add it to cargo
---
As of now, you can insert a beaker, with optional pill bottle, to
produce pills up to 30u.
You can adjust the size of the pills produced with a screwdriver.
The machine can be placed onto tables, or dragged around.

:cl:
add: Adds a hand-powered pill press for ghetto chemistry.
/:cl:

---------

Signed-off-by: GenericDM <34109002+GenericDM@users.noreply.github.com>
Co-authored-by: Generic DM <genericdm.dm@gmail.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>