Makes the exosuit fabricator able to connect to the RND server via multitool (#2336)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
# THE LONG AWAITED CHANGE
## About The Pull Request
You can now finally connect the exosuit fabricator to the RND server
using a multitool, no RND console needed.

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/98485444-98c4-4896-b883-2abbfe22beb0)

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/ab353cea-cccb-4a36-9288-c086a3aacc5d)

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/935aac66-6252-425e-a5dc-f6f148feaed0)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Needing an RND console near an exofab was a leftover thing from days
long past. This is convenient, standardized with other machines, and
much clearer to most players. It even updates automatically instead of
needing to press a button to update it!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: exosuit fabricatiors can now be connected to the RND server via
multitool
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->