Adds a voice log to the wideband (#2307)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a voice log to the wideband that stores the last 50 messages spoken
along with their timestamps to make the wideband more usable without
having to be glued to it

![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/e2897056-4cb2-4530-b6ed-464b7f48157c)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
it's sure to make ships interact more with each other by taking on a
role similar to a chatroom
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a voice log for the wideband
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: BarteG44 <105491762+BarteG44@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>