Make body bag printable in the autolathe (#2285)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Body bags are printable for a price of 2 plastic sheet in every
autolathe/protolathe.

## Why It's Good For The Game

Honestly they are just handy and the design feel basic enough to be
available everywere.

## Changelog

:cl:
add: Body bags in the autolathe.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: ritorizo <ritorizo@localhost>