Tweaks newer rifle projectiles (#2331)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Brings some of the new rifle ammo types since the gun rebalance in line
with the edited ones, and finally removes dismemberment from 300
Blackout.

- 8mm caseless is now brought in line with .308 and 7.62x54mmR
- 300 Magnum does less damage but has the same AP as other rifle
cartridges
- 300 Blackout loses dismember and gains a little AP (same stats as
7.62x39mm)

300 Magnum is the one I'm least sure about here. It's based on 300
Win-Mag, which is a substantially more powerful cartridge than .308 WIN
and similar, so making it just a clone didn't feel right. We might need
to nerf the Smile's magazine and/or cycling rate or something to make up
for this.

## Why It's Good For The Game

consistency within a weapon class is nice

## Changelog

:cl:
balance: nerfed some outlier rifle cartridges for consistency
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->