Removes SSJob (Brutally) (#2127)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

SSJob is a vestigial part of the codebase, quickly bandaged to support
our system. I've slowly stripped it away, but I left it in because it
was rooted deeply in a few pieces of code. Today I finally just decided
to rip that out too, since SSjob is continuing to break things to this
day.

Removes:
- SSjob, replacing it with three global lists
- Job datum slot counts
- Gang gamemode (nonfunctional)
- Revs gamemode (nonfunctional)
- Families gamemode (could you ever call this functional)

Adds:
- Display order support for crew consoles, based on either job datum or
order in the .json file
- Returning someone from ajail returns them to their original ship

## Why It's Good For The Game

SSJob breaks a lot, and just isn't used. This is another piece of global
state that's being removed and replaced with something that better fits
us.

## Changelog

:cl:
del: Gang and revs gamemode (the gear remains)
admin: Removes SSjob, occupations are now stored in a global list
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->