Eyepatch Flopping + Double Eyepatch (#2362)
## About The Pull Request

You can now alt+click any eyepatch (normal, thermal, sechud) to switch
the side that it covers. Examine tells you which side it currently
covers.

You can also add another eyepatch to an eyepatch (only normal because I
don't want to make special ones for the other two types when the first
will probably see the light of day once) to make a DOUBLE EYEPATCH, aka
a blindfold. Interacting with the blindfold in hand gives you the two
eyepatches back.

## Why It's Good For The Game

Switching eyepatch to decide which side you wear it on is just... good,
because character and like, stuff.

Double eyepatch is a full-on meme that makes me giggle a LITTLE but I'm
fine with it not existing.

## Changelog

:cl:
add: Adds the double eyepatch, a blindfold made by adding a normal
eyepatch to another
tweak: Eyepatches can now swap sides with ALT+CLICK
/:cl: