Remaps the Skipper Heavy Cruiser (#2344)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I love the Skipper a lot, but it was getting kind of outdated in terms
of mapping and had a slew of issues that made it harder to enjoy as time
went on. This remap changes the footprint of the Skipper to have a
smaller engineering segment and more room for the crew to move about and
interact.

SDMM:
![2023-09-14 02 49
11](https://github.com/shiptest-ss13/Shiptest/assets/95449138/687cef0a-12bb-475c-abe3-23ebdf4cc538)
The egriegious amounts of wallpipes have been removed, the internal
affairs agent has their own room with privacy shutters to hold meetings,
the cafeteria and crew area have been upgraded and engineering has been
shrunk by about 30%.

Ingame:

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/9b9906be-6782-4718-8c1d-ebab917f689c)
The three engineer slots have been replaced with an atmos tech and an
engineer. A cargo tech slot has been added to manage the cargo office
and handle outpost purchases. There is also now a dedicated cook slot to
occupy the new kitchen and bar.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Ship maintaining. Also wallpipes, wallguns, the injector under a
thruster and various other things aren't good for a map.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Remaps the Skipper Heavy Cruiser
add: Adds an Atmos Tech, Cargo Tech and Cook job slot to the skipper
del: Removes two engineer slots from the skipper.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zevo <95449138+Zevotech@users.noreply.github.com>