Paracelsus Updates: Quality of Life features! (#2370)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR addresses a few things:

- Both Ship and Field Engineer Closets now have prescription mesons for
those who have poor eyesight.
- In the same lieu of helping the visually impaired, a box of
prescription glasses can now be found in the medbay storage area!
- A bottle of universal enzyme has been thrown in for cheese making. Its
important for solarian cuisine!
- A multitude of extra folders now exist around the ship. Documentation
galore!
- The Psychologist's Office Airlock is now properly labeled. Hurray!

And finally,

- This PR fixes the solarian hat. Turns out some commit vaporized the
sprite, so it was left broken for an entire month, whoops.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Gotta maintain my ships yo.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
add: Prescription glasses have been added to the Paracelsus, as well as
the meson variant in the engineering closets!
add: Universal enzyme can now be found in the Paracelsus' freezer!
add: Adds a variety of solarian folders throughout the ship and in the
clerical supplies crate!
fix: The Psychologist's Office Airlock is now properly labeled.
fix: The solarian hat now displays its sprite as it should.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->