Remaps Wasteplanet Pandora Attempt 2 Electric Boogaloo (#2300)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Remaps wasteplanet_pandora to not be a tiny box with a one-way entry
into a locked room with a pandora. The mysterious duplicate lines have
been removed for good this time (I remembered to hit "Sync Changes")

![251871247-0c8241c1-58bf-4659-9218-7edbc9cff921](https://github.com/shiptest-ss13/Shiptest/assets/95449138/59306688-aa99-48bc-8f39-c1d8af2526ac)

![251876112-825e3f5a-4389-4bcd-99cf-49d05e821b4f](https://github.com/shiptest-ss13/Shiptest/assets/95449138/ce8bbfd2-eee6-4cd6-9eb2-401841f53b22)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
A ruin that is just a box with a one way portal into an arena with a
dangerous enemy and no loot just ain't right. Hopefully this map can act
as an example of what to do with megafauna/elite "boss" ruins.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Remapped wasteplanet_pandora.dmm
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zevo <95449138+Zevotech@users.noreply.github.com>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>