Rations 2 Electric Boogalo (#2373)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Fixing all the things wrong with the first ration pr since it got merged
without a test!
If anyone has any issues with rations I'm fixing them here hopefully.

- [x] Better explain ration heaters.
- [x] Make ration heaters send messages to chat about heating.
- [x] Stop rations from acting like beakers and splashing.
- [x] Make powder packs work on reagent containers.
- [x] Make the ration types more clear.
- [x] You are not allowed to munch on powder/liquid
- [x] Make some side dishes heat-able
- [x] Heaters are meant to have more then 1 use
- [x] Ration heater hard deleting

A few other small fixes not worth mentioning

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fixing bugs is a pretty good idea!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: fixed multiple bugs and oversights with rations
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->