Heron- Class! (#1670)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request


![2023 01 19-16 00
15](https://user-images.githubusercontent.com/79304582/213476321-be993cff-b0ac-488b-9c3b-6f33002130b3.png)



![image](https://user-images.githubusercontent.com/79304582/211120744-02f54294-6f8b-4bfd-9904-e6f6df9792d8.png)


This PR adds the Heron-Class Star Dreadnaught (ADMINSPAWN ONLY)!!! It's
a gigantic Nanotrasen Vessel, containing a Dropship, a Mechbay, and a
full Emergency Response Team!

I created this ship as a kind of "opposing force" to the Twinkleshine,
though it would probably heavily differ, in use.
while it has less soldier capacity, less combat gear, and less medical
equipment, it makes up for it with its Marauder mech, Functional"tm"
Supermatter engine, and its state-of-the-art, totally radical, dropship.


the crew consists of:
___
- 1x Captain (Centcom Officer)
- 1x First Officer (hop)
- 1x HOS
- 2x Pilots (for the dropship)
- 1x Security Officer (Internal, actual Security work)
- 6x ERT Officers 
- 1x Mech Pilot
- 1x Engine Technician (Engineers)
- 1x Chief Engineer
- 1x Roboticist
- 2x Medical workers (Paramedic and Medical doctor)
- 1x Atmos tech
- 2x Cargo Crew (quartermaster and cargo tech)
- 1x Cook
- 1x Janitor
____

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Tick the box below (put an X instead of a space between the
brackets) if you have tested your changes and this is ready for review.
Leave unticked if you have yet to test your changes and this is not
ready for review. -->

- [x] I affirm that I have tested all of my proposed changes and that
any issues found during tested have been addressed.

## Why It's Good For The Game
More big ships for admins and players to have fun with is good!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: adds the Heron Class Dreadnaught
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: spockye <79304582+spockye@users.noreply.github.com>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: goober3 <118859017+goober3@users.noreply.github.com>