fixes some issues on the box and li tieguai class ships (#2303)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
-fixed misplaced light switches
-fixed misplaced fire extinguisher cabinets
-got rid of the errors from the li tieg's intern closet
-fixed some areas on both ships
-fixed some aesthetics that got messed up when the floor sprites got
changed
-gave the li tieg an anesthetic closet 
-changed the box bathroom layout for like the 4th time
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
ship maintenance is good i think
## Changelog

:cl:
fix: fixed a few issues with the Box class
fix: fixed a few issues with the Li-tieguai class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: spockye <79304582+spockye@users.noreply.github.com>
Co-authored-by: goober3 <118859017+goober3@users.noreply.github.com>