Removes mentions of Russians, Capitalism/Stalinium Golems, Derelict Drone Subtype, Tweaks AI Ion Laws (#2379)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

It does exactly what it says on the tin.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Not lore compliant. Old sprites with gear that nobody or nothing uses.
"Gas the lizards" is not an acceptable term.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Russian Battle Coat, Russian Uniform, Russian Officer Uniform,
Russian Helmet, Russian Boots, Russian Balaclava
imagedel: Deleted sprites relating to the outfits above.
del: Stalinium and Capitalist Golems
del: Derelict Drones
tweak: The butterbear is now on the Frontiersmen team.
tweak: Modified the list of laws that an AI may recieve from being
ioned.
sounddel: The Soviet Anthem and the US Anthem.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->