Crate shelf (#2374)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This PR intends to add a new method of organizing and storing items in a
way that allows ships to keep more stuff off the floor.

![githubcrate](https://github.com/shiptest-ss13/Shiptest/assets/34109002/4301cb94-907d-40fa-b84a-014b17ae1f29)

Todo:

- [x] Add shelves
- [x] Add the ability for shelves to hold crate
- [x] Add the ability for crates to be removed from shelves
- [x] Arbitrarily tall shelves?
- [x] Add a way to craft them ingame
- [x] Add a way to disassemble them ingame
- [x] Add a check on unloading to ensure they can't be stacked forever
(Kapu's turf.Enter() changes may be ported for this)
- [ ] Fix (or mitigate) the layering problems
- [x] Bring the code up to consistent standards

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I believe that adding more options for organization is a good thing.
Specifically, this will ideally make players more likely to sort items
into crates, rather than just dumping everything in.
I built this with balance in mind, so once finished, there should be no
way to abuse this to get infinitely dense storage, or anything like
that.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Generic DM, PositiveEntropy
add: Shelves! For crates!
fix: Racks can no longer be stacked infinitely.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: GenericDM <34109002+GenericDM@users.noreply.github.com>
Co-authored-by: Imaginos16 <77556824+Imaginos16@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>